Extract HasBusinessSubscription
