+++ app/models/user.rb
@@ -1,11 +1,12 @@
 class User < ApplicationRecord
+  include HasBusinessSubscription
+
   devise :confirmable,
     :database_authenticatable,
     :recoverable,
     :registerable,
     :rememberable,
     :validatable
-  pay_customer
 
   has_many :notifications, as: :recipient, dependent: :destroy
   has_one :business, dependent: :destroy
@@ -19,19 +20,6 @@ class User < ApplicationRecord
   }
 
   scope :admin, -> { where(admin: true) }
-
-  def pay_customer_name
-    business&.name
-  end
-
-  def active_business_subscription?
-    subscriptions.active.any?
-  end
-
-  def active_legacy_business_subscription?
-    legacy_plan = BusinessSubscription::Legacy.new
-    subscriptions.for_name(legacy_plan.name).active.any?
-  end
 
   # Always remember when signing in with Devise.
   def remember_me
